requirements.txt
asgiref==3.2.3
Django==2.1
django-filter==2.2.0
djangorestframework==3.11.0
drf-nested-routers==0.91
pip install psycopg2